Rename cumulative distance files
