+++ src/lib/processing/cumulative-distance.ts
